Added project application control in default voter admin screen

+++ src/main/kotlin/com/terraformation/backend/admin/AdminDefaultVotersController.kt
@@ -1,12 +1,22 @@
 package com.terraformation.backend.admin
 
 import com.terraformation.backend.accelerator.db.DefaultVoterStore
+import com.terraformation.backend.accelerator.db.VoteStore
 import com.terraformation.backend.api.RequireGlobalRole
 import com.terraformation.backend.customer.db.UserStore
 import com.terraformation.backend.customer.model.requirePermissions
+import com.terraformation.backend.db.accelerator.CohortPhase
+import com.terraformation.backend.db.accelerator.VoteOption
+import com.terraformation.backend.db.accelerator.tables.references.COHORTS
+import com.terraformation.backend.db.accelerator.tables.references.PARTICIPANTS
+import com.terraformation.backend.db.accelerator.tables.references.PROJECT_VOTE_DECISIONS
 import com.terraformation.backend.db.default_schema.GlobalRole
+import com.terraformation.backend.db.default_schema.ProjectId
 import com.terraformation.backend.db.default_schema.UserId
+import com.terraformation.backend.db.default_schema.tables.references.PROJECTS
 import com.terraformation.backend.log.perClassLogger
+import org.jooq.DSLContext
+import org.jooq.Record
 import org.springframework.stereotype.Controller
 import org.springframework.ui.Model
 import org.springframework.validation.annotation.Validated
@@ -23,6 +33,8 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes
 class AdminDefaultVotersController(
     private val userStore: UserStore,
     private val defaultVoterStore: DefaultVoterStore,
+    private val dslContext: DSLContext,
+    private val voteStore: VoteStore
 ) {
   private val log = perClassLogger()
 
@@ -31,9 +43,12 @@ class AdminDefaultVotersController(
       model: Model,
       redirectAttributes: RedirectAttributes,
   ): String {
-    requirePermissions { updateGlobalRoles() }
+    requirePermissions { updateDefaultVoters() }
+
+    val projects = getAcceleratorProjects()
 
     model.addAttribute("users", defaultVoterStore.findAll().map { userStore.fetchOneById(it) })
+    model.addAttribute("projects", projects)
 
     return "/admin/defaultVoters"
   }
@@ -41,8 +56,10 @@ class AdminDefaultVotersController(
   @PostMapping("/defaultVoters/add")
   fun addDefaultVoter(
       @RequestParam email: String,
+      @RequestParam projectIds: List<ProjectId>?,
       redirectAttributes: RedirectAttributes,
   ): String {
+    requirePermissions { updateDefaultVoters() }
 
     val userIdForEmail = userStore.fetchByEmail(email)?.userId
     if (userIdForEmail == null) {
@@ -53,6 +70,8 @@ class AdminDefaultVotersController(
     try {
       defaultVoterStore.insert(userIdForEmail)
       redirectAttributes.successMessage = "Default voter added."
+
+      projectIds?.forEach { voteStore.upsert(it, getProjectPhase(it)!!, userIdForEmail) }
     } catch (e: Exception) {
       log.error("Failed to add default voter", e)
       redirectAttributes.failureMessage = "Failed to add default voter: ${e.message}"
@@ -64,11 +83,18 @@ class AdminDefaultVotersController(
   @PostMapping("/defaultVoters/remove")
   fun removeDefaultVoter(
       @RequestParam userId: UserId,
+      @RequestParam projectIds: List<ProjectId>?,
       redirectAttributes: RedirectAttributes,
   ): String {
+    requirePermissions { updateDefaultVoters() }
+
+    val projects = getAcceleratorProjects()
+
     try {
       defaultVoterStore.delete(userId)
       redirectAttributes.successMessage = "Default voter removed."
+
+      projectIds?.forEach { voteStore.upsert(it, getProjectPhase(it)!!, userId) }
     } catch (e: Exception) {
       log.error("Failed to add default voter", e)
       redirectAttributes.failureMessage = "Failed to remove default voter: ${e.message}"
@@ -78,4 +104,49 @@ class AdminDefaultVotersController(
   }
 
   private fun redirectToDefaultVoter() = "redirect:/admin/defaultVoters"
+
+  private fun getAcceleratorProjects(): List<AcceleratorProject> {
+    return dslContext
+        .select(
+            COHORTS.NAME,
+            COHORTS.PHASE_ID,
+            PROJECTS.ID,
+            PROJECTS.NAME,
+            PROJECT_VOTE_DECISIONS.VOTE_OPTION_ID)
+        .from(PROJECTS)
+        .join(PARTICIPANTS)
+        .on(PARTICIPANTS.ID.eq(PROJECTS.PARTICIPANT_ID))
+        .join(COHORTS)
+        .on(PARTICIPANTS.COHORT_ID.eq(COHORTS.ID))
+        .leftJoin(PROJECT_VOTE_DECISIONS)
+        .on(PROJECT_VOTE_DECISIONS.PROJECT_ID.eq(PROJECTS.ID))
+        .orderBy(COHORTS.ID, PROJECTS.ID)
+        .fetch { AcceleratorProject.of(it) }
+  }
+
+  private fun getProjectPhase(projectId: ProjectId): CohortPhase? {
+    return getAcceleratorProjects().firstOrNull { it.projectId == projectId }?.phase
+  }
+}
+
+data class AcceleratorProject(
+    val cohortName: String,
+    val phase: CohortPhase,
+    val projectId: ProjectId,
+    val projectName: String,
+    val voteDecision: VoteOption?
+) {
+  companion object {
+    fun of(
+        record: Record,
+    ): AcceleratorProject {
+      return AcceleratorProject(
+          cohortName = record[COHORTS.NAME]!!,
+          phase = record[COHORTS.PHASE_ID]!!,
+          projectId = record[PROJECTS.ID]!!,
+          projectName = record[PROJECTS.NAME]!!,
+          voteDecision = record[PROJECT_VOTE_DECISIONS.VOTE_OPTION_ID],
+      )
+    }
+  }
 }
